Added marketing Analysis code
A Portuguese banking institution—ran a marketing campaign to convince potential customers to invest in bank term deposit.
Information related to direct marketing campaigns of the bank are as follows.
The marketing campaigns were based on phone calls. Often, the same customer was contacted more than once through phone, in order to assess if they would want to subscribe to the bank term deposit or not.